feat(195): Move spreadsheet code into it's own widget
